+++ jetito/focusspot/focusanalysis.py
@@ -233,22 +233,22 @@ class focusspot_analysis:
                    data_fitted.reshape(self.image_crop.shape[0], self.image_crop.shape[1]),
                    levels=[Z_FWHM], colors=['black'])
 
-        text_legend = (r"$\sigma_x$ = %.1f \u03BCm\n" +
-                       r"$2\sigma_x$ = %.1f \u03BCm\n" +
-                       r"FWHM$_x$ = %.1f \u03BCm\n") % (self.popt_fit[3],
-                                                        2*self.popt_fit[3],
-                                                        2.35*self.popt_fit[3])
+        text_legend = ("$\sigma_x$ = %.1f \u03BCm\n" +
+                       "$2\sigma_x$ = %.1f \u03BCm\n" +
+                       "FWHM$_x$ = %.1f \u03BCm\n") % (self.popt_fit[3],
+                                                       2*self.popt_fit[3],
+                                                       2.35*self.popt_fit[3])
         _ = ax.text(0.07, 0.125, text_legend, horizontalalignment='left', color='white',
                     fontsize=10, weight='bold', verticalalignment='center',
                     transform=ax.transAxes)
 
-        text_legend2 = (r"$\sigma_y$ = %.1f \u03BCm\n" +
-                        r"$2\sigma_y$ = %.1f \u03BCm\n" +
-                        r"FWHM$_y$ = %.1f \u03BCm\n" +
-                        r"q-factor = %.1f %%") % (self.popt_fit[4],
-                                                  2*self.popt_fit[4],
-                                                  2.35*self.popt_fit[4],
-                                                  self.q_factor)
+        text_legend2 = ("$\sigma_y$ = %.1f \u03BCm\n" +
+                        "$2\sigma_y$ = %.1f \u03BCm\n" +
+                        "FWHM$_y$ = %.1f \u03BCm\n" +
+                        "q-factor = %.1f %%") % (self.popt_fit[4],
+                                                 2*self.popt_fit[4],
+                                                 2.35*self.popt_fit[4],
+                                                 self.q_factor)
         _ = ax.text(0.55, 0.125, text_legend2, horizontalalignment='left', color='white',
                     fontsize=10, weight='bold', verticalalignment='center', transform=ax.transAxes)