+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog of Jetito
 
+## v0.4.0
+* Added root mean square (rms) emittance evaluation from pepper pots using experimental data, or GEANT4 simulations.
+  * New examples for the evaluation in the `examples/` folder.
+
 ## v0.3.0
-**Focus spot analysis**
-* Added calculation of focus spot and its parameters from near-field
-* 
+* Added calculation of focus spot and its parameters from near-field.
+* Added electron beam divergence and charge capabilities.
+* Fixed plot problems for focus and farfield analysis, beam charge and divergence.